requirements.txt
openai==0.25.0
python-dotenv==0.21.0
Markdown==3.4.1